Fixed bug causing the whole program to crash
Fixed bug when starting a search with a hard drive added as root search
folder. Hard drives have a path separator included at the end of their
path, which caused the tree view to add another path separator character
when adding a node to the tree, causing an infinite loop because the paths
never match.